improved Queue code
