--- README.md
@@ -1,12 +1,0 @@
-# fp-2023-cub-fall
-
-This is a supplementory repository to the Advanced Functional Programming course taught in the [Constructor University Bremen](https://constructor.university/) by [@danyaberezun](https://github.com/danyaberezun) and [@kajigor](https://github.com/kajigor). 
-
-The syllabus of the course can be found [here](https://docs.google.com/document/d/11nIeQ4LNnz988R2goMwRJFJ2bVCzVCF8/edit?usp=sharing&ouid=106592839861472782542&rtpof=true&sd=true). 
-Please familiarize yourself with it, especially with the criteria for the completion of the course. 
-
-The materials used in lectures, as well as homework assignments, will be published in a separate branch of this repository. 
-Unless stated otherwise, the deadline for a homework assignment is 23:59 of the day preceding the lecture following the homework assignment's publication. 
-To submit the homework assignment, please make a fork of this repository and open a pull-request to the corresponding branch. 
-Please tag the instructor who assigned the homework to ensure that your pull-request has been seen. 
-If your homework assignment has not been checked for a week after submission, please reach out to the instructor on telegram.
