+++ hw01.cabal
@@ -1,0 +1,47 @@
+cabal-version:      3.0
+-- The name of the package.
+name:               hw01
+version:            0.1.0.0
+
+-- The license under which the package is released.
+license:            MIT
+
+-- The file containing the license text.
+license-file:       LICENSE
+
+-- The package author(s).
+author:             razvanperial
+
+-- An email address to which users can send suggestions, bug reports, and patches.
+maintainer:         razvanperial@gmail.com
+
+-- A copyright notice.
+-- copyright:
+category:           Development
+build-type:         Simple
+
+-- Extra doc files to be distributed with the package, such as a CHANGELOG or a README.
+extra-doc-files:    CHANGELOG.md
+
+-- Extra source files to be distributed with the package, such as examples, or a tutorial module.
+-- extra-source-files:
+
+common warnings
+    ghc-options: -Wall
+
+executable hw01
+    -- Import common warning flags.
+    import:           warnings
+
+    -- .hs or .lhs file containing the Main module.
+    main-is:          Main.hs
+
+    -- Other library packages from which modules are imported.
+    build-depends:    base ^>=4.17.0.0, criterion
+    other-modules:    QueueBankers, QueueRealTime, QueuePureFunctional, QueuePhysicist
+
+    -- Directories containing source files.
+    hs-source-dirs:   app, src
+
+    -- Base language which the package is written in.
+    default-language: Haskell2010
